requirements.txt
atari-py==0.1.1
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
gym==0.9.4
idna==2.6
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyglet==1.2.4
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
torch==0.2.0.post3
torchvision==0.1.9
urllib3==1.22
